Added features when player leaves room or leaves the app, the player losses the the other player wins
also added score features, and set android:noHistory=true

+++ app/src/main/java/com/example/guanghuili/checkesandchess/Checkers/Player.java
@@ -57,7 +57,7 @@ public class Player extends RoomManager implements Serializable {
 
     public void updateWinningRate(){
         if((win + loss) != 0){
-            this.winningRate = (double)Math.round((win/(win+loss))*100)/100;
+            this.winningRate = (double)Math.round(((double)win/(double)(win+loss))*100)/100;
         }
     }
